add a method for creating arrays in the variable space
in C one can create an array in the frame of a function using int foo[32];
The allocation is done at compile time so there is no code at runtime. This
functionality was missing in libjit. One was only to create an array using
alloca. This commit adds a function called jit_insn_array that allows allocating
space in the function frame for arrays where the size is known at compile time.