rewrite call.c: there is now a somewhat special ABI for calling a function directly which is used when we know which function we are calling during compile time allowing the caller to check types and set default values. Otherwise a <funcname>.checked function will be called with the old ABI which does type checks and sets default values.
This means in the best case scenario where all types of all parameters of the
caller are known during compile time we have no type checks at all. Additionally
when a value is not provided (i.e. the default value is used) we do that
implicitly and dont have to do a conditional jump